talks_project/requirements.txt
Django==4.0.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn
psycopg2-binary==2.9.9
django-redis
channels_redis
channels
daphne
django-cors-headers
Pillow
django-allauth
social-auth-app-django
pytest
pytest-django
coverage